requirements.txt
tensorflow>=2.13.0
Pillow>=10.0.0
Flask>=2.3.0
Werkzeug>=2.3.0
matplotlib>=3.7.0
numpy>=1.24.0